Separate native and JS prepare code
Create interfaces to separate the prepare functionality in js and native parts.
Ensure better isolation and encapsulation of project preparation process.
