+++ lib/services/platform-service.ts
@@ -23,6 +23,8 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 	private _trackedProjectFilePath: string = null;
 
 	constructor(private $devicesService: Mobile.IDevicesService,
+		private $preparePlatformNativeService: IPreparePlatformNativeService,
+		private $preparePlatformJSService: IPreparePlatformJSService,
 		private $errors: IErrors,
 		private $fs: IFileSystem,
 		private $logger: ILogger,
@@ -30,13 +32,11 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 		private $platformsData: IPlatformsData,
 		private $projectDataService: IProjectDataService,
 		private $hooksService: IHooksService,
-		private $nodeModulesBuilder: INodeModulesBuilder,
 		private $pluginsService: IPluginsService,
 		private $projectFilesManager: IProjectFilesManager,
 		private $mobileHelper: Mobile.IMobileHelper,
 		private $hostInfo: IHostInfo,
 		private $devicePathProvider: IDevicePathProvider,
-		private $xmlValidator: IXmlValidator,
 		private $npm: INodePackageManager,
 		private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		private $projectChangesService: IProjectChangesService,
@@ -141,60 +141,17 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 	private async addPlatformCore(platformData: IPlatformData, frameworkDir: string, platformTemplate: string, projectData: IProjectData, config: IPlatformOptions, nativePrepare?: INativePrepare): Promise<string> {
 		const coreModuleData = this.$fs.readJson(path.join(frameworkDir, "..", "package.json"));
 		const installedVersion = coreModuleData.version;
-		const customTemplateOptions = await this.getPathToPlatformTemplate(platformTemplate, platformData.frameworkPackageName, projectData.projectDir);
-		config.pathToTemplate = customTemplateOptions && customTemplateOptions.pathToTemplate;
+
+		await this.$preparePlatformJSService.addPlatform(platformData, frameworkDir, installedVersion, projectData, config, platformTemplate);
 
 		if (!nativePrepare || !nativePrepare.skipNativePrepare) {
 			const platformDir = path.join(projectData.platformsDir, platformData.normalizedPlatformName.toLowerCase());
 			this.$fs.deleteDirectory(platformDir);
-			await this.addPlatformCoreNative(platformData, frameworkDir, installedVersion, projectData, config);
+			await this.$preparePlatformNativeService.addPlatform(platformData, frameworkDir, installedVersion, projectData, config);
 		}
-
-		const frameworkPackageNameData: any = { version: installedVersion };
-		if (customTemplateOptions) {
-			frameworkPackageNameData.template = customTemplateOptions.selectedTemplate;
-		}
-
-		this.$projectDataService.setNSValue(projectData.projectDir, platformData.frameworkPackageName, frameworkPackageNameData);
 
 		const coreModuleName = coreModuleData.name;
 		return coreModuleName;
-
-	}
-
-	private async addPlatformCoreNative(platformData: IPlatformData, frameworkDir: string, installedVersion: string, projectData: IProjectData, config: IPlatformOptions): Promise<void> {
-		await platformData.platformProjectService.createProject(path.resolve(frameworkDir), installedVersion, projectData, config);
-		platformData.platformProjectService.ensureConfigurationFileInAppResources(projectData);
-		await platformData.platformProjectService.interpolateData(projectData, config);
-		platformData.platformProjectService.afterCreateProject(platformData.projectRoot, projectData);
-	}
-
-	private async getPathToPlatformTemplate(selectedTemplate: string, frameworkPackageName: string, projectDir: string): Promise<{ selectedTemplate: string, pathToTemplate: string }> {
-		if (!selectedTemplate) {
-			// read data from package.json's nativescript key
-			// check the nativescript.tns-<platform>.template value
-			const nativescriptPlatformData = this.$projectDataService.getNSValue(projectDir, frameworkPackageName);
-			selectedTemplate = nativescriptPlatformData && nativescriptPlatformData.template;
-		}
-
-		if (selectedTemplate) {
-			const tempDir = temp.mkdirSync("platform-template");
-			this.$fs.writeJson(path.join(tempDir, constants.PACKAGE_JSON_FILE_NAME), {});
-			try {
-				const npmInstallResult = await this.$npm.install(selectedTemplate, tempDir, {
-					disableNpmInstall: false,
-					frameworkPath: null,
-					ignoreScripts: false
-				});
-				const pathToTemplate = path.join(tempDir, constants.NODE_MODULES_FOLDER_NAME, npmInstallResult.name);
-				return { selectedTemplate, pathToTemplate };
-			} catch (err) {
-				this.$logger.trace("Error while trying to install specified template: ", err);
-				this.$errors.failWithoutHelp(`Unable to install platform template ${selectedTemplate}. Make sure the specified value is valid.`);
-			}
-		}
-
-		return null;
 	}
 
 	public getInstalledPlatforms(projectData: IProjectData): string[] {
@@ -216,8 +173,10 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 	public getPreparedPlatforms(projectData: IProjectData): string[] {
 		return _.filter(this.$platformsData.platformsNames, p => { return this.isPlatformPrepared(p, projectData); });
 	}
+
 	public async preparePlatform(platform: string, appFilesUpdaterOptions: IAppFilesUpdaterOptions, platformTemplate: string, projectData: IProjectData, config: IPlatformOptions, filesToSync?: Array<String>, nativePrepare?: INativePrepare): Promise<boolean> {
 		const platformData = this.$platformsData.getPlatformData(platform, projectData);
+
 		const changesInfo = await this.initialPrepare(platform, platformData, appFilesUpdaterOptions, platformTemplate, projectData, config, nativePrepare);
 		const requiresNativePrepare = (!nativePrepare || !nativePrepare.skipNativePrepare) && changesInfo.nativePlatformStatus === constants.NativePlatformStatus.requiresPrepare;
 
@@ -246,25 +205,6 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 			}
 
 			return valid;
-		}
-	}
-
-	private async cleanProject(platform: string, appFilesUpdaterOptions: IAppFilesUpdaterOptions, platformData: IPlatformData, projectData: IProjectData): Promise<void> {
-		// android build artifacts need to be cleaned up
-		// when switching between debug, release and webpack builds
-		if (platform.toLowerCase() !== "android") {
-			return;
-		}
-
-		const previousPrepareInfo = this.$projectChangesService.getPrepareInfo(platform, projectData);
-		if (!previousPrepareInfo) {
-			return;
-		}
-
-		const { release: previousWasRelease, bundle: previousWasBundle } = previousPrepareInfo;
-		const { release: currentIsRelease, bundle: currentIsBundle } = appFilesUpdaterOptions;
-		if ((previousWasRelease !== currentIsRelease) || (previousWasBundle !== currentIsBundle)) {
-			await platformData.platformProjectService.cleanProject(platformData.projectRoot, projectData);
 		}
 	}
 
@@ -298,10 +238,10 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 
 		const platformData = this.$platformsData.getPlatformData(platform, projectData);
 		const projectFilesConfig = helpers.getProjectFilesConfig({ isReleaseBuild: appFilesUpdaterOptions.release });
-		await this.preparePlatformCoreJS(platform, platformData, appFilesUpdaterOptions, projectData, platformSpecificData, changesInfo, filesToSync, projectFilesConfig);
+		await this.$preparePlatformJSService.preparePlatform(platform, platformData, appFilesUpdaterOptions, projectData, platformSpecificData, changesInfo, filesToSync, projectFilesConfig);
 
 		if (!nativePrepare || !nativePrepare.skipNativePrepare) {
-			await this.preparePlatformCoreNative(platform, platformData, appFilesUpdaterOptions, projectData, platformSpecificData, changesInfo, projectFilesConfig);
+			await this.$preparePlatformNativeService.preparePlatform(platform, platformData, appFilesUpdaterOptions, projectData, platformSpecificData, changesInfo, filesToSync, projectFilesConfig);
 		}
 
 		const directoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
@@ -313,95 +253,6 @@ export class PlatformService extends EventEmitter implements IPlatformService {
 		this.$projectFilesManager.processPlatformSpecificFiles(directoryPath, platform, projectFilesConfig, excludedDirs);
 
 		this.$logger.out(`Project successfully prepared (${platform})`);
-	}
-
-	private async preparePlatformCoreJS(platform: string, platformData: IPlatformData, appFilesUpdaterOptions: IAppFilesUpdaterOptions, projectData: IProjectData, platformSpecificData: IPlatformSpecificData, changesInfo?: IProjectChangesInfo, filesToSync?: Array<String>, projectFilesConfig?: IProjectFilesConfig): Promise<void> {
-		if (!changesInfo || changesInfo.appFilesChanged) {
-			await this.copyAppFiles(platformData, appFilesUpdaterOptions, projectData);
-
-			// remove the App_Resources folder from the app/assets as here we're applying other files changes.
-			const appDestinationDirectoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-			const appResourcesDirectoryPath = path.join(appDestinationDirectoryPath, constants.APP_RESOURCES_FOLDER_NAME);
-			if (this.$fs.exists(appResourcesDirectoryPath)) {
-				this.$fs.deleteDirectory(appResourcesDirectoryPath);
-			}
-		}
-
-		if (!changesInfo || changesInfo.modulesChanged) {
-			await this.copyTnsModules(platform, platformData, projectData, projectFilesConfig);
-		}
-	}
-
-	public async preparePlatformCoreNative(platform: string, platformData: IPlatformData, appFilesUpdaterOptions: IAppFilesUpdaterOptions, projectData: IProjectData, platformSpecificData: IPlatformSpecificData, changesInfo?: IProjectChangesInfo, projectFilesConfig?: IProjectFilesConfig): Promise<void> {
-		if (changesInfo.hasChanges) {
-			await this.cleanProject(platform, appFilesUpdaterOptions, platformData, projectData);
-		}
-
-		if (!changesInfo || changesInfo.changesRequirePrepare) {
-			await this.copyAppFiles(platformData, appFilesUpdaterOptions, projectData);
-			this.copyAppResources(platformData, projectData);
-			await platformData.platformProjectService.prepareProject(projectData, platformSpecificData);
-		}
-
-		if (!changesInfo || changesInfo.modulesChanged || appFilesUpdaterOptions.bundle) {
-			await this.$pluginsService.validate(platformData, projectData);
-
-			const appDestinationDirectoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-			const lastModifiedTime = this.$fs.exists(appDestinationDirectoryPath) ? this.$fs.getFsStats(appDestinationDirectoryPath).mtime : null;
-
-			const tnsModulesDestinationPath = path.join(appDestinationDirectoryPath, constants.TNS_MODULES_FOLDER_NAME);
-			// Process node_modules folder
-			await this.$nodeModulesBuilder.prepareNodeModules(tnsModulesDestinationPath, platform, lastModifiedTime, projectData, projectFilesConfig);
-		}
-
-		if (!changesInfo || changesInfo.configChanged || changesInfo.modulesChanged) {
-			await platformData.platformProjectService.processConfigurationFilesFromAppResources(appFilesUpdaterOptions.release, projectData);
-		}
-
-		platformData.platformProjectService.interpolateConfigurationFile(projectData, platformSpecificData);
-		this.$projectChangesService.setNativePlatformStatus(platform, projectData,
-			{ nativePlatformStatus: constants.NativePlatformStatus.alreadyPrepared });
-	}
-
-	private async copyAppFiles(platformData: IPlatformData, appFilesUpdaterOptions: IAppFilesUpdaterOptions, projectData: IProjectData): Promise<void> {
-		platformData.platformProjectService.ensureConfigurationFileInAppResources(projectData);
-		const appDestinationDirectoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-
-		// Copy app folder to native project
-		this.$fs.ensureDirectoryExists(appDestinationDirectoryPath);
-		const appSourceDirectoryPath = path.join(projectData.projectDir, constants.APP_FOLDER_NAME);
-
-		const appUpdater = new AppFilesUpdater(appSourceDirectoryPath, appDestinationDirectoryPath, appFilesUpdaterOptions, this.$fs);
-		appUpdater.updateApp(sourceFiles => {
-			this.$xmlValidator.validateXmlFiles(sourceFiles);
-		});
-	}
-
-	private copyAppResources(platformData: IPlatformData, projectData: IProjectData): void {
-		const appDestinationDirectoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-		const appResourcesDirectoryPath = path.join(appDestinationDirectoryPath, constants.APP_RESOURCES_FOLDER_NAME);
-		if (this.$fs.exists(appResourcesDirectoryPath)) {
-			platformData.platformProjectService.prepareAppResources(appResourcesDirectoryPath, projectData);
-			const appResourcesDestination = platformData.platformProjectService.getAppResourcesDestinationDirectoryPath(projectData);
-			this.$fs.ensureDirectoryExists(appResourcesDestination);
-			shell.cp("-Rf", path.join(appResourcesDirectoryPath, platformData.normalizedPlatformName, "*"), appResourcesDestination);
-			this.$fs.deleteDirectory(appResourcesDirectoryPath);
-		}
-	}
-
-	private async copyTnsModules(platform: string, platformData: IPlatformData, projectData: IProjectData, projectFilesConfig?: IProjectFilesConfig): Promise<void> {
-		const appDestinationDirectoryPath = path.join(platformData.appDestinationDirectoryPath, constants.APP_FOLDER_NAME);
-		const lastModifiedTime = this.$fs.exists(appDestinationDirectoryPath) ? this.$fs.getFsStats(appDestinationDirectoryPath).mtime : null;
-
-		try {
-			const tnsModulesDestinationPath = path.join(appDestinationDirectoryPath, constants.TNS_MODULES_FOLDER_NAME);
-			// Process node_modules folder
-			await this.$nodeModulesBuilder.prepareJSNodeModules(tnsModulesDestinationPath, platform, lastModifiedTime, projectData, projectFilesConfig);
-		} catch (error) {
-			this.$logger.debug(error);
-			shell.rm("-rf", appDestinationDirectoryPath);
-			this.$errors.failWithoutHelp(`Processing node_modules failed. ${error}`);
-		}
 	}
 
 	public async shouldBuild(platform: string, projectData: IProjectData, buildConfig: IBuildConfig, outputPath?: string): Promise<boolean> {
